get refactor working again
