--- packages/wdio-browser-runner/src/browser/reporter.ts
@@ -1,10 +1,0 @@
-// @ts-ignore
-const BaseReporter = Mocha.reporters.html
-
-export default class HTMLReporter extends BaseReporter {
-    constructor (runner: any, options: any) {
-        super(runner, options)
-    }
-
-    addCodeToggle () {}
-}
