+++ packages/wdio-browser-runner/src/vite/server.ts
@@ -10,7 +10,8 @@ import type { Services } from '@wdio/types'
 
 import { testrunner } from './plugins/testrunner.js'
 import { userfriendlyImport } from './utils.js'
-import { PRESET_DEPENDENCIES, DEFAULT_VITE_CONFIG, MESSAGE_TYPES } from './constants.js'
+import { PRESET_DEPENDENCIES, DEFAULT_VITE_CONFIG } from './constants.js'
+import { MESSAGE_TYPES } from '../constants.js'
 import type { ConsoleEvent, HookTriggerEvent, CommandRequestEvent, CommandResponseEvent, SocketMessage, HookResultEvent } from './types'
 
 import { BROWSER_POOL, SESSIONS } from '../constants.js'
