+++ packages/wdio-browser-runner/src/vite/utils.ts
@@ -1,4 +1,63 @@
-import type { FrameworkPreset } from '../types'
+import fs from 'node:fs/promises'
+import path from 'node:path'
+
+import { createRequire } from 'node:module'
+
+import type { Environment, FrameworkPreset } from '../types'
+
+export async function getTemplate (options: WebdriverIO.BrowserRunnerOptions, env: Environment, spec: string) {
+    const root = options.rootDir || process.cwd()
+    const require = createRequire(path.join(root, 'node_modules'))
+
+    let vueDeps = ''
+    if (options.preset === 'vue') {
+        try {
+            const vueDir = path.dirname(require.resolve('vue'))
+            const vueScript = (await fs.readFile(path.join(vueDir, 'dist', 'vue.global.prod.js'), 'utf-8')).toString()
+            vueDeps += /*html*/`
+            <script type="module">
+                ${vueScript}
+                window.Vue = Vue
+            </script>`
+            const vueCompilerDir = path.dirname(require.resolve('@vue/compiler-dom'))
+            const vueCompilerScript = (await fs.readFile(path.join(vueCompilerDir, 'dist', 'compiler-dom.global.prod.js'), 'utf-8')).toString()
+            vueDeps += /*html*/`
+            <script type="module">
+                ${vueCompilerScript}
+                window.VueCompilerDOM = VueCompilerDOM
+            </script>`
+        } catch (err: any) {
+            throw new Error(
+                `Fail to set-up Vue environment: ${err.message}\n\n` +
+                'Make sure you have "vue" and "@vue/compiler-dom" installed as dependencies!\n' +
+                `Error: ${err.stack}`
+            )
+        }
+    }
+
+    return /* html */`
+    <!doctype html>
+    <html>
+        <head>
+            <title>WebdriverIO Browser Test</title>
+            <link rel="icon" type="image/x-icon" href="https://webdriver.io/img/favicon.png">
+            <link rel="stylesheet" href="https://unpkg.com/mocha@10.0.0/mocha.css">
+            <script type="module">
+                /**
+                 * Inject environment variables
+                 */
+                window.__wdioEnv__ = ${JSON.stringify(env)}
+                window.__wdioSpec__ = '${spec}'
+                window.__wdioEvents__ = []
+            </script>
+            ${vueDeps}
+        </head>
+        <body>
+            <div id="mocha"></div>
+            <script async type="module" src="@wdio/browser-runner/setup"></script>
+        </body>
+    </html>`
+}
 
 export async function userfriendlyImport (preset: FrameworkPreset, pkg?: string) {
     if (!pkg) {
@@ -13,4 +72,16 @@ export async function userfriendlyImport (preset: FrameworkPreset, pkg?: string)
             `Please run:\n\n\tnpm install ${pkg}\n\tor\n\tyarn add --dev ${pkg}`
         )
     }
+}
+
+export async function getErrorTemplate (filename: string, error: Error) {
+    return /*html*/`
+        <pre>${error.stack}</pre>
+        <script type="module">
+            window.__wdioErrors__ = [{
+                filename: "${filename}",
+                message: \`${error.message}\`
+            }]
+        </script>
+    `
 }
